Reformatted Environment as Gym environment
