+++ flowing_basin/analyses/simulation_vs_history.py
@@ -1,5 +1,5 @@
 from flowing_basin.tools import RiverBasin
-from flowing_basin.solvers.rl import Training
+from flowing_basin.solvers.rl import Environment
 from cornflow_client.core.tools import load_json
 import pandas as pd
 import numpy as np
@@ -8,7 +8,7 @@ from matplotlib import pyplot as plt
 import os
 
 # Path and extension in which to save the generated graphs
-SAVE_PLOT = True
+SAVE_PLOT = False
 PATH_START = "simulation_vs_history_graphs/sim_v_hist_NO-NA_"
 PATH_END = ""
 EXTENSION = ".png"
@@ -48,7 +48,7 @@ START = datetime.strptime("2021-09-29 01:00", "%Y-%m-%d %H:%M")
 path_constants = "../data/rl_training_data/constants.json"
 path_training_data = "../data/rl_training_data/training_data_NO_NA.pickle"
 df = pd.read_pickle(path_training_data)
-instance = Training.create_instance(
+instance = Environment.create_instance(
     length_episodes=LENGTH_EPISODE,
     constants=load_json(path_constants),
     training_data=df,
