+++ flowing_basin/solvers/rl/__init__.py
@@ -1,3 +1,3 @@
 from .environment import Environment
-from .agent import Agent
-from .training import Training
+from .rl_train import RLTrain
+from .rl_run import RLRun
